- Änderungen aus anderem Projekt übernommen

+++ ALR_ROB_2.0/src/alrrob/robot/Robot.java
@@ -1,7 +1,7 @@
 package alrrob.robot;
 
 public interface Robot {
-	public void moveRobotForMS(int ms) throws Exception;
+	public void moveRobotForMS(double ms) throws Exception;
 //	public void moveRobot();
 //	public void stopRobot();
 	public void rotateRobot(double angle) throws Exception;
